django-shapes/requirements.txt

Brotli==1.0.9
cffi==1.15.0

dj-database-url==0.5.0


django-multiselectfield==0.1.12
django-select-multiple-field==0.4.2

gunicorn==20.1.0
html5lib==1.1


pycparser==2.21
pydyf==0.1.2

reportlab==3.6.7
six==1.16.0
sqlparse==0.4.2
tinycss2==1.1.1
weasyprint==54.2
webencodings==0.5.1
whitenoise
zopfli
